+++ neural_modelling/src/neuron/synapses.h
@@ -45,6 +45,7 @@ static inline index_t synapses_get_ring_buffer_index_combined(
 }
 
 // Converts a weight stored in a synapse row to an input
+// UNUSED!!!!!! CHECK convert_rate_to_input in synapses.c!!!
 static inline input_t synapses_convert_weight_to_input(
         uint32_t weight, uint32_t left_shift) {
     union {
@@ -71,7 +72,7 @@ bool synapses_initialise(
         address_t synapse_params_address, address_t direct_matrix_address,
         uint32_t n_neurons, uint32_t n_synapse_types,
         uint32_t **ring_buffer_to_input_buffer_left_shifts,
-        address_t *direct_synapses_address, uint32_t starting_rate);
+        address_t *direct_synapses_address, REAL starting_rate);
 
 void synapses_do_timestep_update(timer_t time);
